requirements.txt
colorama
python-dotenv
PyGithub
requests
groq
IPython
jinja2
gensim
transformers
torch